requirements.txt
absl-py==1.0.0
albumentations==1.1.0
alembic==1.7.7
alpaca-ml==0.9.2
anyio==3.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.5
attrs==21.2.0
autopage==0.5.0
Babel==2.9.1
backcall==0.2.0
black==22.3.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.4
colorlog==6.6.0
commentjson==0.9.0
commonmark==0.9.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
deepdiff==5.8.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dotmap==1.3.26
dppy==0.3.2
entrypoints==0.3
filelock==3.4.0
fonttools==4.28.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.1
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lark-parser==0.7.8
lazy-object-proxy==1.7.1
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
msgpack==1.0.3
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
opencv-python-headless==4.5.5.62
optuna==2.10.0
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pbr==5.9.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
prettytable==3.3.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydocstyle==6.1.1
Pygments==2.10.0
pylint==2.13.9
pyparsing==3.0.6
pyperclip==1.8.2
pyrsistent==0.18.0
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==2.0.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
qudida==0.0.4
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
rich==11.0.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.36
stevedore==3.5.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==2.0.1
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.6.0
